KEYCLOAK-7193 Correct base path in product build pom (#122)
The PR for KEYCLOAK-6909 (#121) was modified to put the new stuff in a
subdirectory. I omitted to check that the pom would actually build from this
new location, and it doesn't.

In a couple places I've added a `../` so that the paths point to the repo root
again, so the build actually works now. Sorry for not checking this before.